Feature/playlist ksh

See merge request s12-bigdata-recom-sub1/S12P21C104!116

+++ frontend/app/src/main/java/com/whistlehub/common/view/navigation/AppContentNavGraph.kt
@@ -70,8 +70,8 @@ fun AppContentNavGraph(
                 navController
             )
         }
-        composable(route = Screen.Profile.route + "/{memberId}"){ backStackEntry ->
-        val memberId = backStackEntry.arguments?.getString("memberId")
+        composable(route = Screen.Profile.route + "/{memberId}") { backStackEntry ->
+            val memberId = backStackEntry.arguments?.getString("memberId")
             ProfileScreen(
                 memberId?.toInt() ?: -1,
                 paddingValues = paddingValues,
@@ -108,7 +108,7 @@ fun AppContentNavGraph(
             if (playlistId != null) {
                 PlaylistTrackListScreen(
                     paddingValues,
-                    playlistId.toInt(),
+                    playlistId,
                     navController,
                     trackPlayViewModel
                 )
